+++ api/controllers/recipients/add-computers.js
@@ -14,7 +14,7 @@ module.exports = {
       if (!this.req.isSocket) { return exits.error(new Error('This controller requires a websocket.')) }
 
       // Add the recipient
-      var label = await sails.helpers.recipients.add(sails.sockets.getId(this.req), this.req.payload.host, 'computers', this.req.payload.host)
+      var response = await sails.helpers.recipients.add(sails.sockets.getId(this.req), this.req.payload.host, 'computers', this.req.payload.host)
 
       // Subscribe to websockets dealing with remote audio calls for DJ Controls
       if (this.req.isSocket) {
@@ -24,7 +24,7 @@ module.exports = {
       }
 
       // Return the host label object
-      return exits.success({ label: label })
+      return exits.success(response)
     } catch (e) {
       return exits.error(e)
     }